Reduce number of line pragmas
With this change only the required line pragmas are generated. Instead
of

  <original line>
  {-# LINE ... #-}
  <original line>
  {-# LINE ... #-}
  <original line>

the pragmas are only introduced around the lines generated from the
`require` directives:

  module Foo where
  require Bar

is turned into

  module Foo where
  import Bar (Bar)
  {-# LINE ... #-}
  import qualified ...

The main reason for this change is that the previous behaviour didn't
play well with haddock comments.

  -- | First line
  -- Second line

would get turned into

  -- | First line
  {-# LINE ... #-}
  -- Second line

which lead to the second line not being included in the haddock output.