Filter autorequire imports based on the host module name
The host module is the module on which we are currently working. All
require directives in that file are taken as they are because we assume
the user knows what he is doing.

The require directives which come from the Requires file are only taken
when the required module is different from the host module.

If no `module .. where` directive is seen, no filtering occurs.